at before mt
annual income tax is needed in the calculation for monthly income tax. putting the formula for manual income tax first confuses the program since the calculation for at has not been done yet, that's why it tends to say that at has not yet been declared, which in return, does not ;et the program finish.